Move InventoryGUIScreen into src/, include/ files
