+++ game/project.hpp
@@ -8,6 +8,7 @@
 #include "logging.hpp"
 
 
+
 class Project : public UIScreen
 {
 public:
