+++ game/script_helper.hpp
@@ -1,4 +1,7 @@
 
+#include <inventory_gui/notification.hpp>
+
+
 
 class ScriptHelper
 {
